move the DiagLayerContainer class into its own file
it used to be co-located with DiagLayer...
